requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
jsonpickle==3.2.1
python-logging-loki @ git+https://github.com/waylayio/python-logging-loki.git@69428186d64a560eda20701f0910b86bedf0ebcf
requests==2.32.3
rfc3339==6.2
urllib3==2.2.1
